Cleans up some ordering in os_compat.c.
This rearranges the order of some conditionals and includes, in order
to avoid some gratuitous includes where unnecessary.

Due to the fact that some actual definitions for the clock_gettime
fallback are embedded directly in gpsd_config.h, it's necessary to
include <time.h> prior to including it.  This will be fixed once the
relevant definitions are moved to the upcoming os_compat.h.

This also replaces the "UNUSED" in clock_gettime with a void cast,
eliminating the need for compiler.h.

TESTED:
Ran "scons build-all check" on OSX 10.9, OSX 10.12 (which has real
clock_gettime) and Ubuntu 14.  Also verified that the clock_gettime
fallback is present in OSX 10.9, but absent in OSX 10.12 and in Ubuntu
14.