Centralizes daemon() calls.
This defines a new function os_daemon() (in os_compat.c), which is
either the old replacement daemon() renamed, or a wrapper around the
actual daemon() call.  This allows any issues related to daemon()
(which exist on some platforms) to be dealt with in one place.  No
such changes are present yet, so platforms giving warnings for the use
of daemon() continue to do so, but now only in the compilation of
os_compat.c.  Unfortunately, the current build procedure typically
compiles os_compat.c multiple times, so the warnings still appear
multiple times.

TESTED:
Ran "scons build-all check" on OSX 10.9, OSX 10.12, Ubuntu 14, and
FreeBSD 10.3.