Fixes some issues with C++11.
In compiler.h, it adds the missing "std::" namespace prefixes to the
memory_barrier() definition.

In gpsd.h it:

1) Moves the include of compiler.h outside the conditional 'extern
"C"', since the "atomic" stuff in the former is incompatible with the
latter.

2) Fixes DEVICEHOOKPATH for C++11 (C++11 requires spaces between
literals and string macros).

3) Cleans up some ordering of system includes left over from the
former head/tail setup.

TESTED:
Ran "scons build-all check" on OSX 10.5-10.12, Ubuntu 14, CentOS 7,
Fedora 25, FreeBSD 10.3, OpenBSD 5.6 (32- and 64-bit), and NetBSD
6.1.5.  Also tested Qt builds with OSX 10.9 (Qt4 and Qt5), OSX 10.12
(Qt5), and Fedora 25 (Qt5).
Observed the correct DEVICEHOOKPATH in the log on OSX.