Rereverts "Work around yet another scons fuckup."
Since the original problem is still not reproducible, and since Debian
compatibility is now tested with Travis, this patch is again being
removed.

TESTED:
In the previous incarnation, ran "scons build-all check" on four
versions of OSX and two versions of Linux, as well as FreeBSD,
OpenBSD, and NetBSD.  Retested now on OSX 10.9 and Fedora 25.