Windows libgps support
Add Windows versions for network functions and disable unused features.

Add detection of various headers and functions that aren't available in
Windows. Note that netlib_localsocket() has no functional implementation
on Windows, but it isn't to be used on that platform and it's not part of
libgps anyway.

Using send() rather than write() seems to work on Windows.

For Windows need to ensure networking is initialised on opening of sockets
(and then correspondingly shutdown upon closing).

Note that within gpsd.h the termios structures and serial related functions
are removed from the Windows build. These are only accessed in serial.c by
gpsd, so since the Windows build is only generating libgps there is
currently no need to modify serial.c.

And for os_compat.h, daemon() is simply disabled under Windows as it's not
used within libgps.

TESTED:
Confirmed compiles under a cross compiler.
Manual build and run of test_libgps on Windows which successfully connects
to a host running GPSD
Manual build DLL version and link with a Windows version of a program
(Viking) that then successfully connects and monitors positions from GPSD

Otherwise no effect on current supported systems.
'scons build-all check' - passes.

Signed-off-by: Fred Wright <fw@fwright.net>