Cleans up os_daemon() in Windows case.
This provides an always-failing os_daemon() for platforms (e.g.,
Windows) that can't provide a real daemon() call.  This is more
consistent than simply omitting the function.

It also bases the condition on the lack of fork(), rather than the
presence of winsock2.h.

TESTED:
Ran "scons build-all check" on OSX, Linux, FreeBSD, OpenBSD, and
NetBSD.  Verified that HAVE_FORK is set in all cases.