Ensure gpxlogger doesn't crash on bad filename
Need to ensure logfile is reset to 'stdout' when the specified output file
is not available (e.g. no permission to write it) otherwise it is continued
to be used when it is NULL and bad things happen.

TESTED:
Running as a normal user it doesn't crash with a bad filename:
./gpxlogger -f /root/CannotWriteHere.gpx

Signed-off-by: Fred Wright <fw@fwright.net>