Fixes broken build of www/hardware.html.
Commit 8200880a (polyglot rework) broke the build of hardware.html,
due to problems with default encodings in gpscap.py.  This went
unnoticed for nearly a year since the 'www' target is not part of the
default build, and because the failure didn't propagate to an overall
build failure.

Since the intended encoding for both the input and output is UTF-8,
this is fixed by specifying that encoding explicitly.  This needs to be
done in two places:

1) The invocation of the ConfigParser's read() method needs to specify
the encoding of the input.  This argument isn't available in Python
2.6, so there's a fallback for that case (where it's unneeded).

2) The encoding of the piped output needs to be specified.  This is
done by setting the PYTHONIOENCODING environment variable when running
gpscap.py.

TESTED:
Ran "scons build-all www check".  Verified that the output looks
plausible in a browser, including the two umlauted characters.
Also verified that running gpscap.py standalone (with PYTHONIOENCODING
set) works correctly with all supported Python versions.