Fixes error detection on build of www/hardware.html.
The invocation of gpscap.py was ignoring the exit status, thereby
allowing the build to succeed even when the program didn't.  The fix
is simply to chain the shell commands with '&&' rather than ';'.

TESTED:
Verified that the known bug now causes a build failure.
Also verified that "scons build-all check" still works.