Fixes broken 'www' build with missing asciidoc.
Although there's logic to build a partial 'www' target when asciidoc
is missing, this was failing due to the inappropriate (and normally
redundant) inclusion of installation.html in the htmlpages list.
Removing the entry fixes this.

TESTED:
Building 'www' without asciidoc now succeeds (with the usual warning
message), and building with asciidoc still succeeds and includes
installation.html in the results.