Suppresses manbuilder chatter in silent mode.
Neither xsltproc nor xmlto has an option to suppress the "Note:
Writing ..." messages.  This adds a "filtered spawn" function which
suppresses stderr output when the exit code is 0, and applies it to
the manpage building when running in silent mode.

It also removes the unused 'mangenerator' variable.

TESTED:
Verified that manbuilder chatter is gone with "scons -s build-all",
but still present without the "-s".
Verified that the error is reported in "-s" mode when the XML input is
garbled.
Ran "scons build-all www check".