Updates packet-regress test data.
This updates the test data to reflect the two new tests that were
added in commit 58a35243.  This was neglected before due to a bug
(commit 40c40cc4) that causes the packet-regress test to be
inaccessible.  This test still doesn't pass due to a bug introduced in
commit fbaaaa76, but this update makes the current failure equivalent
to the previous failure.  In particular, it avoids gratuitous
miscompares due to mismatches in the line numbers.

TESTED:
Running the test manually now gets the same errors as it did prior to
the addition of the new tests.