Fixes disabled packet-regress test.
The method-regress test was misnamed packet-regress, which overrode
the real packet-regress test, causing it to be disabled.  Now that
packet-regress's previously undetected failures have been fixed, it
can be reenabled.

TESTED:
Ran "scons build-all www check", and observed the inclusion of the
missing test (with no errors).
Also ran it with manually garbled test data, and observed the failure.