Fixes broken 'GPSD' SHM export.
Commit 22c3faf added an IPC_RMID operation right after the shmat(), in
order to keep the segment from lingering after gpsd exits.  But this
clears the key, making it impossible for a client to attach to the
segment.

This fix moves the IPC_RMID to shm_release(), so that it's not invoked
until gpsd is done with the segment.  This required saving the shmid
in the gps_context structure.

TESTED:
On both OSX and Linux, used ipcs to verify that the segment is present
with the correct ID while gpsd is running, and disappears after gpsd
exits.  Also verified that warning message for shmctl() appears when
two gpsds conflict over the SHM segment.