Fixes gpscat raw-mode file-input termination.
When reading from a file in packetizer mode, gpscat correctly
terminates at EOF.  However, in raw mode it simply hung in a loop
retrying the read.  This change makes it exit on EOF.

This change also removes the useless buffer append logic in raw mode,
which was a leftover from the old code that attempted to break on
newlines (removed by commit e4cbd2daf).

TESTED:
Ran gpscat in raw mode on all test/daemon log files, and observed
proper termination.  Tested with Python 2.7 and Python 3.5.