Disables Travis builds of C++ code with clang.
There is a compatibility issue with some versions of clang++ and some
C++ headers, related to the __float128 type, which causes builds to
fail.  Until this is reproducible outside Travis, it's best to disable
the C++ components when building with clang, rather than having the
test fail for reasons having nothing to do with GPSD itself.
Non-clang C++ builds are still tested as are non-C++ builds with
clang.

TESTED:
Script change to accept added options work.
Real test won't happen until this code is pushed.
Only Travis-related files are touched.