Fixes broken regression test data for uBlox NEO-M8N.
Commit 71a487d8, "Support UBX NAV-PVT", added a new regression test
case, but the .chk data didn't actually match.  All mismatches were
a difference of 0.01 in the 'epc' value.  Whether the difference
is due to an actual code difference or excessive numeric sensitivity
remains to be seen, but in the meantime this gets rid of the failure
which seems to be consistent.

TESTED:
Ran "scons build-all check", which now succeeds.