Fixes SConstruct for SCons 3.0.0.
SCons 3.0.0 introduced a bug where the print_funtion future import is
inflicted on the SConstruct script, making 'print' as a statement
illegal.  This is expected to be fixed in SCons 3.0.1, but in the
meantime it's necessary to switch to the print_function mode for
compatibility.  Fortunately, there were only three print statements in
the whole file.

This is not a complete Python 3 fix; it simply restores correct
operation when running SCons under Python 2.

TESTED:
Arraged to test all three print() instances under OSX.
Tested "scons build-all check" under OSX, Ubuntu, CentOS, Fedora,
FreeBSD, OpenBSD, and NetBSD.