gpspipe improvement: add option to exit after delay
If the GPS receiver cease to supply data to gpsd, gpspipe hangs
indefinitely waiting for gpsd.

This can happen if the GPS is turned off by external means, e.g. by an
AT command on the command port of a cellular card.

In order to use gpspipe to get a fix within bounded time, this patch
adds an option to specify a timout, so that I can do things like:
   $x=$(gpspipe -x 5 -w|sed -n '/TPV/{p;q}')

And get either the first TPV or an empty string if no fix is available.

Signed-off-by: Gary E. Miller <gem@rellim.com>