gpsd.c: fix gpsd master/slave mode
The gpsd command parser was splitting the ? from commands (WATCH, POLL,
etc.), and then sending the ? to the gpsd master in a separate packet.

But the gpsd is not buffering network data, so failing to recognize
The two packets of ? and WATCH as one complete command.  Adding full
buffering to the gpsd master looks a lot harder than just sending
the ?WATCH in one packet.  Also leading to some code simplification
in the gpsd slave.  Been broken since 2013 at least.

Many thanks to Virgin Orbit for the support finding and fixing this.