pylintrc: Disable many annoying pylint messages
gpsd does not try to conform to pylint, but pylint can be useful
to find things like unused variables, missing docstrings, undefined
variables, unused imports, etc.

The whitespace checker in pylint conflicts with that in pycodestyle.
pycodestyle replaces pep8 for checking PEP8 conformity.

Use pylint as an advisory tool, not as a blunt instrument.