Disables xgps[speed] -V in python-versions check.
Some of the Gtk imports used by xgps and xgpsspeed aren't merely
imports, but also actually establish a connection to the X11 server.
Not only does this require that a usable X11 server be available (in
some cases launching one), but it may also take significant time even
when it works.  It would be best to fix xgps and xgpsspeed to avoid
those imports for options such as -V and -h, but in the meantime it's
undesirable to include those programs in the version check.  This fix
replaces their invocations with a comment.

The change that introduced the version checks almost certainly would
have failed Bernd's Travis CI checks, if that mechanism hadn't died
166 commits ago.

TESTED:
Both "scons python-versions" and "scons check" now produce the
expected results without launching an X11 server or delaying.