Adds test for xgps dependencies.
This adds a new test_xgps_deps.py, which duplicates the imports
needed by xgps and xgpsspeed, as a quick way to verify that the
proper prerequisites are in place.  It avoids the imports that
require (and connect to) an available X11 server.

TESTED:
Ran successfully with proper dependencies, and verified failure
with either cairo or gobject3 removed.