Includes xgps deps test in test list.
This adds test-xgps-deps to the test-nondaemon list (and hence to the
overall 'check' list).

Since the Travis setup doesn't include the necessary prerequisites,
it forces xgps=no for Travis.  At the time this code was originally
tested, Travis was still working.

TESTED:
Verified that the test is run, and that the test series fails iff
one of the dependencies is broken.