Fixes possible X11 launch in python-versions test.
Some of the Gtk imports may have the side effect of launching an
X11 server if one isn't currently available (at least on the Mac).
This can be suppressed by clearing the DISPLAY environment variable
prior to those imports.  This fix arranges to apply that to the
environment used by the python-versions test.

TESTED:
Ran "scons python-versions" on OSX, and observed no X11 launch.