Align COPYRIGHT files with Project License
Most of the files are tagged SPDX BSD-2-Clause.  Emails on the
gpsd-dev mailing list also confirm that the intention of the lead
developer, ESR, is that the BSD-2-Clause licence be used.

I believe there is consensus on this.

This patch:
1. Makes some minor formatting changes (tabs -> spaces)
2. Removes pseudo HTML formatting
3. Removes the mistaken 3rd clause included (non-endorsement)
4. Adds numbers to the clause, to follow the text on the OSI site
5. Replaces the word "Regents" with "Copyright Holder" in the
   disclaimer, as per the OSI text.

Please add Sign-Offs if I have captured the consensus correctly.

Signed-off-by: Sanjeev Gupta <ghane0@gmail.com>
Signed-off-by: Gary E. Miller <gem@rellim.com>
Signed-off-by: Chris Kuethe <chris.kuethe@gmail.com>
Signed-off-by: Gerry Creager - NOAA Affiliate <gerry.creager@noaa.gov>
Signed-off-by: Fred Wright <fw@fwright.net>