Fixes ubxtool and zerk in direct-serial mode.
The attempted workaround for the writeTimeout->write_timeout change in
serial.Serial() didn't work, since including the unexpected
"alternate" keyword argument caused an error.  This was observed with
both V2.7 and V3.0 of pySerial, so it's not clear how this code ever
worked for anyone.  It now uses only the correct argument name.

There are two other deprecated items, inWaiting and flushInput,
but these have not yet been removed in V3.0, so those are not
yet fixed (other than the comments).

TESTED:
Ran "ubxtool -p VER ..." against a u-Blox LEA-M8T in direct-serial
mode, with pySerial both 2.7 and 3.0.  Did not test zerk, due to the
lack of appropriate hardware, but its changes are identical to those
in ubxtool.