Make sure 'python-versions' uses the right Python.
Although in-tree testing is sometimes able to find the Python
libraries via program-relative addressing, later Python versions name
compiled extensions in a version-dependent manner.  This means that
programs need to run with the specific version of Python that the
extensions were compiled for.  This change invokes the "target_python"
explicitly for the version checks, rather than using the
system-default Python.  This is the same approach that is taken by all
other Python-based tests.

TESTED:
The 'python-versions' test now passes with "target_python=python3.7",
while the system-default Python is 2.7.  This is after already
building the extensions; another commit will add the missing
dependencies.