test_mktime: Fix build on platforms needing librt.
On some older plaforms, clock_gettime() is contained in librt, thus
requiring linking with it.  A mechanism already exists for doing this
where needed, but it was missing from the recipe for test_mktime.

TESTED:
On a BeagleBone running Debian 7, test_mktime now builds successfully.
It still needs another fix for platforms with a 32-bit time_t.