Fixes time-regress for 32-bit time_t platforms.
One of the test cases for the unix<->iso8601 tests uses a value
which exceeds the time range on systems with a 32-bit time_t.  This
adds a check to skip such cases (with a warning message) on such
platforms.

TESTED:
Ran time-regress on both a 64-bit time_t machine and a 32-bit time_t
machine.  It now succeeds in both cases, with warnings in the
32-bit case.