Fixes symlink handling in zip target.
By default, zip follows symlinks rather than storing them as links in
the archive.  This resulted in redundant copies of the 'gps'
subdirectory, which not only wasted space, but also wouldn't actually
work in practice since they wouldn't get populated with the build
products.  Adding the 'y' option to the zip invocation fixes this.

TESTED:
Verified that the resulting zip archive has the expected symlinks, and
that the content matches the repo.