Fixes unreliability in CheckCompilerOption().
The code to restore the old CFLAGS when an option test failed was
sometimes a victim of Python's "list by reference rather than value"
issue.  The effect was that the removal of the failing option
sometimes didn't work.  The bug was only observed when the build was
being run by MacPorts, for some unknown reason.  This has apparently
been broken ever since the introduction of CheckCompilerOption(),
though it didn't cause trouble until "Wvla" was added to the option
list, where it broke builds on OSX 10.5 and 10.6, whose compilers
don't support the option.

The fix is simply to use the "[:]" construct to copy the list.

TESTED:
Tested both "bare" build and MacPorts "destroot" on MacPro 10.9, and
VMs for macOS 10.5-10.13.  Failures on 10.5 and 10.6 are now gone.