gpsfake: Allow timeout to be specified.
This adds an argument to allow overriding the default 60-second
timeout on gpsd's activity.  A value of 0 disables the timeout
completely.  Note that the timeout mechanism is somewhat sloppy, and
typically adds about 20s to the value.

TESTED:
Ran with unspecified (default) timeout, a short timeout, a long
timeout, and no timeout, all with reasonable results.