gpsfake: Add option for debugging with lldb.
This adds the -G option, which is like the -g option but using
lldb rather than gdb.  It's not quite equivalent, since lldb
doesn't have the "-tui" option, although some versions of lldb
have a gui mode that can be entered via command.

TESTED:
Verified that -G launches with lldb as expected.