gpsfake: Disable timeout when debugging.
Since debugging typically involves stopping at breakpoints, the
timeout tends to get in the way.  This adds an implied -W0 to the -g
and -G options to avoid that.  If for some reason the timeout is
actually wanted, it can be reenabled by providing a -W option after
the -g or -G.

TESTED:
Verified that -G disables the timeout.  Did not test -g, though the
code is the same for it.