gpsfake: Notice when daemon dies.
Death of the daemon (after initial connection) wasn't being noticed,
so that the daemon dying (e.g., by quitting it from a debugger) left
gpsfake looping and spewing output.  This puts an alive check in the
main loop, as well as in an inner loop that's sometimes relevant.

TESTED:
Quit the daemon after stopping at breakpoints in a couple of different
cases; gpsfake now quits with an error.  It does still spew a bunch of
duplicated output in some cases, though.