GREIS: Fixes raw-data test failure on some platforms.
This was observed on OSX 10.5, both x86 and PPC, though with different
exact symptoms.  It could have appeared on any platform where
"opportunistic initialization" was absent.  The problem is that some
of the raw.meas fields are never populated at all by the GREIS driver,
but are still converted to JSON, with not-entirely-predictable
results, and hence test failures.  In the x86 case, "snr" was the
offender, while on PPC it was "freqid" and "lli".  This fix
initializes all the fields of the 'meas' substructure that are
otherwise unwritten by this driver.  It did not change the output in
the previously passing cases.

TESTED:
All tests now pass on OSX 10.5-10.13, including 10.5 PPC.