gpsutils: Fix unix_to_iso8601 NaN case.
Since the return type of unix_to_iso8601 is non-const, it's illegal
to return a pointer to a string constant.  Some compilers report
a warning for this.  The fix is to copy the string to the provided
buffer.

It would be nice to add a test case for this, but there doesn't seem
to be any portable way to get NaN as a constant in C.

TESTED:
The warning is gone, and the tests still pass.