Fix shared library installs on OSX.
Shared libraries on OSX need to have their IDs set to their full paths
in order to get the proper paths into executables.  Ideally, SCons
would take care of this, but doesn't.  It's necessary to use the OSX
install_name_tool on an installed library to fix that.

This adds code to do that, verifying that install_name_tool actually
exists, and falling back to installing without it (the former
behavior) otherwise.  However, the CheckProg function didn't exist in
some old versions of SCons, so if it's missing (on OSX), the code just
assumes that the tool is present and hopes for the best.  The version
of SCons provided by MacPorts is known to have CheckProg.

TESTED:
Ran installs on OSX 10.9, 10.13, and 10.5 (PPC), as well as Ubuntu
14.04.  Verified that the tool is used and has the intended effect
when present on OSX.  Verified that the install works without using it
on OSX without it, as well as Linux without it, including when a dummy
version is present on Linux.  Did *not* test with an old SCons.