ubxtool: Adds interface option for UBX-CFG-PRT.
The UBX-CFG-PRT command can specify any port, though it defaults to
the current port.  This option allows the port ID to be specified,
either as a name or as an integer.  At present, it only affects the
PRT preset, though it could be used for other port-related commands in
the future.

Note that the baud rate is only applicable to the UART port, so
there's no point in applying this to the -S option.

TESTED:
Tested on LEA-6S, LEA-M8F, LEA-M8T, and	LEA-M8N