ubxtool: Reworks -c option so that it's usable.
It appears that the -c code was more or less inherited from zerk and
essentially useless for the UBX message format.  This reworks it so
that the argument is a comma-separated list of hex bytes giving the
class, ID, and any payload.  The header, length, and checksum are
added automatically, making it reasonably convenient to specify
arbitrary commands.  The only limitation is the need to specify
multibyte quantities as individual bytes.

TESTED:
Tested on LEA-6S, LEA-M8F, LEA-M8T, and	LEA-M8N