ubxtool: Adds commas to raw payload output.
This both improves the readability of the raw dump, and formats it
such that it can be fed back to a subsequent -c command, making
incremental settings tweaks easier.  Unfortunately, verbosity 5 is
needed to see this; the verbosity meanings are not very well
organized.  Filtering the output with grep is helpful.

TESTED:
Tested on LEA-6S, LEA-M8F, LEA-M8T, and	LEA-M8N