ubxtool: Fixes SAVE/RESET/CFG-CFG bugs.
The masks for SAVE and RESET were not defined as intended by the
comments, including having one of the bits in the wrong byte.

Two conditionals (only one actually effective) were checking the
value of 'type' rather than the 'save_clear' argument, probably
due to an incomplete rename.  This rendered SAVE inoperative,
and due to the implied load, it had the effect of resetting the
parameters to their saved values (usually the defaults, due to
the broken SAVE).

The decode for CFG-CFG had similar bugs related to bit positions.

TESTED:
Tested on LEA-6S, LEA-M8F, LEA-M8T, and	LEA-M8N