ubxtool/zerk: Makes PySerial a "soft" dependency.
This allows the absence of the PySerial ('serial') module, as long as
the -f option is not used.  This not only avoids the PySerial
dependency for informational options like -h and -V, but also allows
full functionality in the "via gpsd" mode.

The 'gps_io' class was taking the 'serial' module (mistakenly referred
to as a class as an init argument, but wasn't making any use of that
flexibility.  Since fully supporting that would complicate the error
message, it's simply removed instead.

The zerk changes are untested, due to the lack of appropriate
hardware.  However, they are identical to the changes in the
corresponding code in ubxtool, which is tested.

The manpages are updated to mention the pyserial dependency for -f.

The install of ubxtool and zerk is no longer conditional on PySerial.

TESTED:
Tested ubxtool with PySerial 2.7, 3.0, 3.4, and none at all, verifying
that -f works with all three versions, and other modes work without.
Also tested with Python 3.4-3.7 (PySerial 3.4 only).