ubxtool: Allow UART selection for speed setting.
This allows the -i option to specify the port for the -S option.  It
currently defaults to port 1, the first or only UART port.  It would
be better (as noted in the FIXME) to default to the current port, but
that would require implementing a mechanism to determine that.

TESTED:
On a NEO-M8T, tested various -i values (including none).  Proper
behavior was observed, though for some reason this receiver accepts
CFG-PRT for the nonexistent port 2, though it rejects polling port 2.