driver_nmea0183: Enforce NMEA 4.0 type PRNs.
Up until now, the PWN was whatever the GPS said it was.  With NMEA 4.10
this is no longer sustainable.  The u-blox 9 defaults to NMEA 4.10, sort
of.  It sometimes puts NMEA 4.0 PRNs into NMEA 4.10 sentences.

Some people like to use both u-blox binary and NMEA sentences mixed
together, so the time has come to enforce the PRN complies with
NMEA 4.0 extended numbering.  Sadly this can not map exactly into
u-blox numbering, but is a start.

Fixing this uncovered other examples of non NMEA compliant sentences in
the regressions.  So many regressions changed.  The most obvious
change is that NMEA 4.0 puts the SBAS PRN in the range 33-64 and
152-158, not 120-158 as u-blox does.