SConstruct: Add environment variables to be preserved
The build is set to pass only environment variables listed in
SConstruct, which breaks various attempts to pass environment
variables to children of the build (such as compilers).  There was
already a list of variables to be preserved.  This commit reorganizes
the list by purpose to make it easier to understand.

Add three variables to the list: CWRAPPERS_CONFIG_DIR, used by pkgsrc,
and CCACHE_DIR and CCACHE_RECACHE, used by ccache.