serial.c: Extend local cfmakeraw to Solaris
POSIX does not specify cfmakeraw (but it is very common).  gpsd
provides an implementation, previously used only on cygwin.  Also use
this implementation on Solaris, and explain the issue more clearly,
with cross-references between the implementation and declaration.

Extend the comment about defining NAN for Solaris (C99 does not
require NAN to be defined, and it is not clear that this is still
required).