gpsd: Change version from 3.19-dev to 3.19~dev
While gpsd other than releases should not be released as packages, it
is helpful to test the release/packaging chain at intermediate points
in development.  There is not agreement in the packaging world or in
individual packages about whether - separates words within a package
name, or within a package version.  Both Debian and pkgsrc say that
versions should not contain '-'.  Avoid needlessly contributing to
this confusion by labeling the development version 3.19~dev instead, a
form that has historically been used.

Add comment in SConstruct near the version declaration that explains
the rationale for ~.