Fix the handling of defaults from the environment.
The way whis was handled before had various issues:

- scons' naming scheme for compiler flags is different from what we know
from autotools and (afaik) all other build systems. As it seems to
filter flags (or they are listed in the wrong environment variable and
not used at the appropriate time), most of them get lost. So far I have
not seen builds to fail becase of that.
- to make sure flags end up in the appropriate setting, use
env.MergeFlags for *FLAGS. This works well for me, but might break
special CXX/CCFLAGS settings. So far I was not able to find a better way
to handle this.
- the lower/original MergeFlags part did not work at all as
  for flag in ....:
    if i in os.eviron:
does not depend on the flag to be checked, but on the last checked flag
from the loop before, where i was used. As TAR is usually not in the
environment, this code was propably always dead.
- the LD/SHLINK switch was done the wrong way around and was never able
to work at all.

tl;dr: the number of newly broken things is for sure less than the
number of things that actually never worked before.

Similar issues seems to exist in the part of the code where Python build
configs are merged into the python env....