SConstruct: Fix C++ build for timespec_str.
This new source needed to be added to the list of sources that are
compiled as C, rather than C++, when building the C++ library.
Otherwise, the function is sometimes not found at link time, probably
due to name mangling issues.

TESTED:
Ran "build-all check" on MacPro/10.9, MacPro/10.14, MacPro/10.15,
MacBookPro/10.9, and VMs for OSX 10.5-10.13, Ubuntu 14.04, CentOS 7,
Fedora 25, FreeBSD 10.3, OpenBSD 5.6 (32- and 64-bit), and NetBSD
6.1.5.  Ran "build-all test-quick" on PowerBook/10.5, to avoid one
regression test failure (to be fixed later).
CI tests also pass (except FreeBSD, which lacks a runner).