SConstruct: Fix test_[Q]gpsmm builds on OpenBSD.
Commits 66763ce0e1 and 689ae13200 were the wrong approach to fixing
the C++ timespec_str problem, taking the rather strange approach of
linking with both the static and dynamic libraries.  The real problem
was the naming issue caused by compiling timespec_str as C++ rather
than C, which is now fixed (05da457e7).  On OpenBSD, this was
interacting with another bug causing the gpsmm tests to fail to link
against the dynamic library, and hence causing build failures.

This change does not fix the bug (which may be in SCons) where the
unversioned library symlinks don't get created, but it avoids the
failing case here and permits the tests to link as they did before.

TESTED:
Ran "build-all check" on MacPro/10.9, MacPro/10.14, MacPro/10.15,
MacBookPro/10.9, and VMs for OSX 10.5-10.13, Ubuntu 14.04, CentOS 7,
Fedora 25, FreeBSD 10.3, OpenBSD 5.6 (32- and 64-bit), and NetBSD
6.1.5.  Ran "build-all test-quick" on PowerBook/10.5, to avoid one
regression test failure (to be fixed later).
CI tests also pass (except FreeBSD, which lacks a runner).