SConstruct: Fix potential build crash with some compilers.
The code that removes C-only flags from qt_flags doesn't allow for the
possibility that those flags may not be present.  This would result in
an exception from the remove().  Wrapping the latter with a try/except
fixes that.

Note that this problem hasn't actually been observed in practice, and
is simply being proactively fixed.

TESTED:
Ran "build-all check" on MacPro/10.9, MacPro/10.14, MacPro/10.15,
MacBookPro/10.9, and VMs for OSX 10.5-10.13, Ubuntu 14.04, CentOS 7,
Fedora 25, FreeBSD 10.3, OpenBSD 5.6 (32- and 64-bit), and NetBSD
6.1.5.  Ran "build-all test-quick" on PowerBook/10.5, to avoid one
regression test failure (to be fixed later).
CI tests also pass (except FreeBSD, which lacks a runner).