fake.py: Eval is evil
I don't think there is a need to be able to specify python code in the
WRITE_PAD variable anymore.

Before d42f3f9d1a770a6adcbfd3d7a232206268fc2bd0 it was only possible to
specify changes to the pre-defined value in the environment, but this
limitation is gone.

So better get rif of the eval.