fake: Set WRITE_PAD to 0.0 for NetBSD
A recent change to use TCP instead of pty for regression testing --
appearing sensible and harmless -- has apparently changed timing, and
with 0.004 (on one computer) fails, while 0.01 works.

Because we are late in the release cycle, just bump up WRITE_PAD for
NetBSD, and hope that soon after release the actual problem will be
found.