gpsprof: Cope with TPV statements without fixes
At some point, gpsd started providing TPV json objects for times when
no fix was computed.  This is useful, as it conveys that the receiver
was operating, as opposed to not having data.  However, data consumers
must be careful to process these differently.  gpsprof, in particular,
assumed without checking that e.g. lat/lon were present in a TPV
message, leading to failure.

This commit simply skips TPV objects that have a fix mode of no fix.
In addition, it adds an explicit exception if the set of fixes ends up
empty, rather than throwing mysterious errors about array bounds.