SConstruct: Use -f with rm for files that might not be writable
SConstruct copies python files to tmp.py, compiles them, and then
removes them.  However, cp is documented to preserve modes of files
(modified by umask), and rm of a file that is not writable is
documented to prompt for confirmation.  (I don't understand how this
works anywhere, but that's not important -- the fix is easy and
harmless.)

https://pubs.opengroup.org/onlinepubs/9699919799/utilities/cp.html
https://pubs.opengroup.org/onlinepubs/9699919799/functions/open.html
https://pubs.opengroup.org/onlinepubs/9699919799/utilities/rm.html