SConstruct: Use target_python when compiling programs in scons check
Before, the python version used by scons was used, which is unrelated
to the desired python for this build of gpsd.

With this change, sys.executable is no longer used by SConstruct,
other than as a default value of target_python if that is not set.

This resolves a failure to build/check when using
target_python=python3.7 when scons uses python2.7.